// types/HomePage.d.ts
export type FaqItem = {
    question: string;
    answer: string;
    defaultOpen?: boolean;
  };
  
  export type FeatureItem = {
    title: string;
    description: string;
    href: string;
    imageUrl: string;
  };
  export type NewsHighlightItem = {
    tag: string;
    title: string;
    description: string;
    imageUrl: string;
    href: string;
  };
  
  